feat: enhance seed phrase protection by excluding extension pages
